+++ JuiceMaker/JuiceMaker/View/Label/FruitStockLabel.swift
@@ -1,0 +1,14 @@
+//
+//  FruitStockLabel.swift
+//  JuiceMaker
+//
+//  Created by hugh, yeton on 2022/05/06.
+//
+
+import UIKit
+
+class FruitStockLabel: UILabel {
+    func convertToFruit() -> Fruit? {
+        return nil
+    }
+}
