+++ JuiceMaker/JuiceMaker/View/Label/MangoLabel.swift
@@ -1,0 +1,14 @@
+//
+//  MangoLabel.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class MangoLabel: FruitStockLabel {
+    override func convertToFruit() -> Fruit? {
+        return .mango
+    }
+}
