+++ JuiceMaker/JuiceMaker/View/Label/PineappleLabel.swift
@@ -1,0 +1,14 @@
+//
+//  PineappleLabel.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class PineappleLabel: FruitStockLabel {
+    override func convertToFruit() -> Fruit? {
+        return .pineapple
+    }
+}
