+++ JuiceMaker/JuiceMaker/View/Label/StrawberryLabel.swift
@@ -1,0 +1,14 @@
+//
+//  StrawberryLabel.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class StrawberryLabel: FruitStockLabel {
+    override func convertToFruit() -> Fruit? {
+        return .strawberry
+    }
+}
