+++ JuiceMaker/JuiceMaker/View/Stepper/BananaStepper.swift
@@ -1,0 +1,14 @@
+//
+//  BananaStepper.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class BananaStepper: EditStockStepper {
+    override func convertToFruit() -> Fruit? {
+        return .banana
+    }
+}
