+++ JuiceMaker/JuiceMaker/View/Stepper/EditStockStepper.swift
@@ -1,0 +1,17 @@
+//
+//  EditStockStepper.swift
+//  JuiceMaker
+//
+//  Created by hugh, yeton on 2022/05/10.
+//
+
+import UIKit
+
+class EditStockStepper: UIStepper {
+    func convertToFruit() -> Fruit? {
+        return nil
+    }
+}
+
+
+
