+++ JuiceMaker/JuiceMaker/View/Stepper/KiwiStepper.swift
@@ -1,0 +1,14 @@
+//
+//  KiwiStepper.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class KiwiStepper: EditStockStepper {
+    override func convertToFruit() -> Fruit? {
+        return .kiwi
+    }
+}
