+++ JuiceMaker/JuiceMaker/View/Stepper/MangoStepper.swift
@@ -1,0 +1,14 @@
+//
+//  MangoStepper.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class MangoStepper: EditStockStepper {
+    override func convertToFruit() -> Fruit? {
+        return .mango
+    }
+}
