+++ JuiceMaker/JuiceMaker/View/Stepper/PineappleStepper.swift
@@ -1,0 +1,14 @@
+//
+//  PineappleStepper.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class PineappleStepper: EditStockStepper {
+    override func convertToFruit() -> Fruit? {
+        return .pineapple
+    }
+}
