+++ JuiceMaker/JuiceMaker/View/Stepper/StrawberryStepper.swift
@@ -1,0 +1,14 @@
+//
+//  StrawberryStepper.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/05/13.
+//
+
+import UIKit
+
+class StrawberryStepper: EditStockStepper {
+    override func convertToFruit() -> Fruit? {
+        return .strawberry
+    }
+}
